backend/pyproject.toml
[project]
name = "contactgraph-backend"
version = "1.0.0"
description = "ContactGraph backend API"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "google-auth==2.23.4",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.1.1",
    "google-api-python-client==2.108.0",
    "python-dotenv==1.0.0",
    "pydantic>=2.6.0",
    "pytest==7.4.3",
    "httpx==0.25.2",
]
requires-python = ">=3.10"
